Added backend pages
CHANGES MADE:
(All Pages) - Slight change in the wording for login box and links to
'add event' page.
(database) - The SQL file added is for creating the database for testing
use.
(Events & EventDetails) - Created these two PHP pages. Events.php is the
catalog of upcoming events. EventDetails is the custom generated page
for an individual event. **These pages are functional but NEED some
fixing on the interface**
(CSS) - Nothing is removed from the original CSS, just more code is at
the very bottom. This was added while trying to format the interface of
Events and EventDetails.php
(Login/outAction & AdminMode) - Created these three PHP pages.
LoginAction serves as a checker when an admin attempts to login. It
verifies the username and password match. If verified, the admin enters
AdminMode.php. This page is empty now, but is planned to be a checklist
form that displays all PENDING events and allows the admin to approve or
delete each one. LogoutAction goes to use when an admin clicks the
“logout” button, the session is ended and they’re taken back to the
homepage. **These pages are still in development. I just wanted them to
exist so that if anyone else has time before I get back to it can help
out**